Fixed lines on hover, broken image, deleted excess css
In the last revision hovering over an image in the gallery would make vertical and horizontal lines appear over it. What I was doing was scaling the normal image up to a bigger size. By having the initial scale of the image be smaller than normal and then scaling it back up to normal, it fixed the issue. YEAHH!!! Anyway, I also fixed one broken image, which was a live image that wouldn't display right on some browsers. I also removed excess CSS code to clean it up a bit.
Also changed the homepage message.